refactor: rename flow output_fn to on_done
