+++ tests/integration/issues/github_1072/test_queryset.py
@@ -42,10 +42,10 @@ def test_queryset_with_struct(random_workspace, mocker):
     response_mock_label2_docs = mocker.Mock(wraps=validate_label2_docs)
     with f:
         # keep all the docs
-        f.index(docs, output_fn=response_mock_all_docs, callback_on='body')
+        f.index(docs, on_done=response_mock_all_docs, callback_on='body')
         # keep only the docs with label2
         qs = QueryLang(FilterQL(priority=1, lookups={'tags__label': 'label2'}, traversal_paths=['r']))
-        f.index(docs, queryset=qs, output_fn=response_mock_label2_docs, callback_on='body')
+        f.index(docs, queryset=qs, on_done=response_mock_label2_docs, callback_on='body')
 
     response_mock_all_docs.assert_called()
     response_mock_label2_docs.assert_called()
