+++ tests/integration/mime/test_segmenter.py
@@ -22,7 +22,7 @@ def test_dummy_seg(mocker):
     response_mock = mocker.Mock(wraps=validate)
     f = Flow().add(uses='DummySegment')
     with f:
-        f.index(input_fn=random_docs(10, chunks_per_doc=0), output_fn=response_mock)
+        f.index(input_fn=random_docs(10, chunks_per_doc=0), on_done=response_mock)
     response_mock.assert_called()
 
 
@@ -30,7 +30,7 @@ def test_dummy_seg_random(mocker):
     response_mock = mocker.Mock(wraps=validate)
     f = Flow().add(uses=os.path.join(cur_dir, '../../unit/yaml/dummy-seg-random.yml'))
     with f:
-        f.index(input_fn=random_docs(10, chunks_per_doc=0), output_fn=response_mock)
+        f.index(input_fn=random_docs(10, chunks_per_doc=0), on_done=response_mock)
     response_mock.assert_called()
 
 
@@ -38,5 +38,5 @@ def test_dummy_seg_not_random(mocker):
     response_mock = mocker.Mock(wraps=validate)
     f = Flow().add(uses=os.path.join(cur_dir, '../../unit/yaml/dummy-seg-not-random.yml'))
     with f:
-        f.index(input_fn=random_docs(10, chunks_per_doc=0), output_fn=response_mock)
+        f.index(input_fn=random_docs(10, chunks_per_doc=0), on_done=response_mock)
     response_mock.assert_called()
