+++ tests/unit/drivers/test_concat_driver.py
@@ -63,6 +63,6 @@ def test_concat_embed_driver(mocker):
             .join(needs=['a', 'b'], uses='- !ConcatEmbedDriver | {}'))
 
     with flow:
-        flow.index(input_fn=input_fn, output_fn=response_mock, callback_on='body')
+        flow.index(input_fn=input_fn, on_done=response_mock, callback_on='body')
 
     response_mock.assert_called()
